pyproject.toml
[project]
name = "s3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.39.2",
    "python-dotenv>=1.1.1",
]
